requirements.txt
google-generativeai
python-dotenv
requests
fastapi
uvicorn
